+++ src/integrations/mod.rs
@@ -1,0 +1,5 @@
+//! Provides extra integrations for seamless use with common libraries and frameworks.
+
+#[cfg(feature = "axum")]
+pub mod axum;
+pub mod vercel_aisdk_ui;